+++ web/src/cache.rs
@@ -34,7 +34,7 @@ pub(crate) enum MutationValue {
     TributeDeleted(String),
     GameFilled(String),
     NewTributes(Vec<Tribute>),
-    TributeUpdated(String, u8),
+    TributeUpdated(String, u32),
 }
 
 #[derive(PartialEq, Debug)]
